fix int to long to allow for more than 4 billion events

+++ src/NeXusDataset/NeXusDataset.h
@@ -63,7 +63,7 @@ public:
 
 /// \brief Represents the index register for searching a large NXlog
 /// (relatively) quickly based on timestamp.
-class CueIndex : public ExtensibleDataset<std::uint32_t> {
+class CueIndex : public ExtensibleDataset<std::uint64_t> {
 public:
   CueIndex() = default;
   /// \brief Create the cue_index dataset of NXLog and NXevent_data.
@@ -122,7 +122,7 @@ public:
 
 /// \brief Represents the (radiation) detector event index that ties
 /// EventTimeZero to event id and offset in a NXevent_data.
-class EventIndex : public ExtensibleDataset<std::uint32_t> {
+class EventIndex : public ExtensibleDataset<std::uint64_t> {
 public:
   EventIndex() = default;
   /// \brief Create the event_index dataset of NXevent_data.
